refact: move :platform-register's rule to :platform-register-test
