+++ build.gradle
@@ -11,6 +11,7 @@ buildscript {
     dependencies {
         classpath _plugins.gradle
         classpath _plugins.kotlin
+        classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.7.20'
     }
 }
