+++ platform-register/src/androidTest/java/com/jeanbarrossilva/memento/platform/register/RoomRegisterTests.kt
@@ -1,6 +1,6 @@
 package com.jeanbarrossilva.memento.platform.register
 
-import com.jeanbarrossilva.memento.platform.register.rule.RegisterTestRule
+import com.jeanbarrossilva.memento.platform.register.test.RegisterTestRule
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert.assertNotNull
