+++ platform-register/src/main/java/com/jeanbarrossilva/memento/platform/register/RoomRepository.kt
@@ -6,7 +6,7 @@ import com.jeanbarrossilva.memento.core.register.infra.Repository
 import com.jeanbarrossilva.memento.platform.register.note.NoteDao
 import com.jeanbarrossilva.memento.platform.register.note.NoteEntity
 
-internal class RoomRepository(private val dao: NoteDao) : Repository {
+class RoomRepository(private val dao: NoteDao) : Repository {
     override suspend fun getNotes(): Map<Path, List<Note>> {
         return dao.selectAll().map(NoteEntity::toNote).groupBy {
             it.path
