+++ settings.gradle
@@ -5,3 +5,4 @@ include ':feature-notes'
 include ':feature-editor'
 include ':core-register'
 include ':platform-register'
+include ':platform-register-test'
