+++ versions.gradle
@@ -35,6 +35,9 @@ buildscript {
             room: '2.5.0',
             test: [
                 core: '1.5.0',
+                ext: [
+                    junit: '1.1.5'
+                ],
                 runner: '1.5.2'
             ]
         ]
